bird2: improve protocol equality
One way to look at it is that one would expect that an element without a suffix would not require a specific protocol and therefore all protocol specific elements should have a suffix.

Another point of view, is that it is simply not an equal treatment of protocols and subconsciously also gives the impression that IPv6 is not the default but an afterthought.

I think in general one could also go less in the direction of equal treatment and more in the direction of labelling IPv4 as a legacy protocol.
Only the impracticality of something like `ip` and `iplegacy` in everyday use keeps me from proposing something like that.